Corrections after PR
