--- requiremenets.txt
@@ -1,10 +1,0 @@
-Django==1.11.13
-django-debug-toolbar==1.9.1
-factory-boy==2.11.1
-Faker==0.8.16
-ipaddress==1.0.22
-python-dateutil==2.7.3
-pytz==2018.4
-six==1.11.0
-sqlparse==0.2.4
-text-unidecode==1.2
